requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
feedparser==6.0.11
frozenlist==1.6.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
jiter==0.10.0
multidict==6.4.4
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.81.0
pandas==2.2.2
pillow==11.0.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.67.1
twilio==9.6.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.20.0
